+++ bungee/src/main/java/com/jannik_kuehn/loritimebungee/BungeePluginMessenger.java
@@ -1,6 +1,7 @@
 package com.jannik_kuehn.loritimebungee;
 
 import com.jannik_kuehn.common.api.common.CommonSender;
+import com.jannik_kuehn.common.api.logger.LoriTimeLogger;
 import com.jannik_kuehn.common.module.messaging.PluginMessaging;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.connection.Server;
@@ -14,9 +15,12 @@ import java.util.UUID;
 public class BungeePluginMessenger extends PluginMessaging implements Listener {
     private final LoriTimeBungee loriTimeBungee;
 
+    private final LoriTimeLogger log;
+
     public BungeePluginMessenger(final LoriTimeBungee loriTimeBungee) {
         super(loriTimeBungee.getPlugin());
         this.loriTimeBungee = loriTimeBungee;
+        this.log = loriTimeBungee.getPlugin().getLoggerFactory().create(BungeePluginMessenger.class);
     }
 
     @EventHandler
@@ -25,7 +29,7 @@ public class BungeePluginMessenger extends PluginMessaging implements Listener {
             return;
         }
         if (event.getSender() instanceof final ProxiedPlayer player) {
-            loriTimePlugin.getLogger().severe(player.getName() + " tried to change the plugin message of LoriTime!");
+            log.error(player.getName() + " tried to change the plugin message of LoriTime!");
             return;
         }
         if (event.isCancelled()) {