+++ bungee/src/main/java/com/jannik_kuehn/loritimebungee/listener/PlayerNameBungeeListener.java
@@ -1,6 +1,7 @@
 package com.jannik_kuehn.loritimebungee.listener;
 
 import com.jannik_kuehn.common.LoriTimePlugin;
+import com.jannik_kuehn.common.api.logger.LoriTimeLogger;
 import net.md_5.bungee.api.event.PostLoginEvent;
 import net.md_5.bungee.api.plugin.Listener;
 import net.md_5.bungee.event.EventHandler;
@@ -11,8 +12,11 @@ public class PlayerNameBungeeListener implements Listener {
 
     private final LoriTimePlugin plugin;
 
+    private final LoriTimeLogger log;
+
     public PlayerNameBungeeListener(final LoriTimePlugin plugin) {
         this.plugin = plugin;
+        this.log = plugin.getLoggerFactory().create(PlayerNameBungeeListener.class);
     }
 
     @EventHandler
@@ -23,7 +27,7 @@ public class PlayerNameBungeeListener implements Listener {
             try {
                 plugin.getNameStorage().setEntry(uuid, name, true);
             } catch (final Exception ex) {
-                plugin.getLogger().warning("could not save player name and uuid " + name, ex);
+                log.warn("could not save player name and uuid " + name, ex);
             }
         });
     }