+++ bungee/src/main/java/com/jannik_kuehn/loritimebungee/listener/TimeAccumulatorBungeeListener.java
@@ -1,6 +1,7 @@
 package com.jannik_kuehn.loritimebungee.listener;
 
 import com.jannik_kuehn.common.LoriTimePlugin;
+import com.jannik_kuehn.common.api.logger.LoriTimeLogger;
 import com.jannik_kuehn.common.exception.StorageException;
 import net.md_5.bungee.api.event.PlayerDisconnectEvent;
 import net.md_5.bungee.api.event.PostLoginEvent;
@@ -13,8 +14,11 @@ public class TimeAccumulatorBungeeListener implements Listener {
 
     private final LoriTimePlugin plugin;
 
+    private final LoriTimeLogger log;
+
     public TimeAccumulatorBungeeListener(final LoriTimePlugin plugin) {
         this.plugin = plugin;
+        this.log = plugin.getLoggerFactory().create(TimeAccumulatorBungeeListener.class);
     }
 
     @EventHandler
@@ -25,7 +29,7 @@ public class TimeAccumulatorBungeeListener implements Listener {
             try {
                 plugin.getTimeStorage().startAccumulating(uuid, now);
             } catch (final StorageException ex) {
-                plugin.getLogger().warning("could not start accumulating online time for player " + uuid, ex);
+                log.warn("could not start accumulating online time for player " + uuid, ex);
             }
         });
     }
@@ -38,7 +42,7 @@ public class TimeAccumulatorBungeeListener implements Listener {
             try {
                 plugin.getTimeStorage().stopAccumulatingAndSaveOnlineTime(uuid, now);
             } catch (final StorageException ex) {
-                plugin.getLogger().warning("error while stopping accumulation of online time for player " + uuid, ex);
+                log.warn("error while stopping accumulation of online time for player " + uuid, ex);
             }
         });
     }