+++ common/src/main/java/com/jannik_kuehn/common/module/afk/AfkStatusProvider.java
@@ -2,6 +2,7 @@ package com.jannik_kuehn.common.module.afk;
 
 import com.jannik_kuehn.common.LoriTimePlugin;
 import com.jannik_kuehn.common.api.LoriTimePlayer;
+import com.jannik_kuehn.common.api.logger.LoriTimeLogger;
 import com.jannik_kuehn.common.api.scheduler.PluginTask;
 
 import java.util.HashMap;
@@ -10,7 +11,9 @@ import java.util.OptionalLong;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class AfkStatusProvider {
-    private final LoriTimePlugin plugin;
+    private final LoriTimePlugin loriTimePlugin;
+
+    private final LoriTimeLogger log;
 
     private final ConcurrentHashMap<LoriTimePlayer, Long> afkCheckedPlayers;
 
@@ -21,7 +24,8 @@ public class AfkStatusProvider {
     private long afkConfigTime;
 
     public AfkStatusProvider(final LoriTimePlugin loriTimePlugin, final AfkHandling afkHandling) {
-        this.plugin = loriTimePlugin;
+        this.loriTimePlugin = loriTimePlugin;
+        this.log = loriTimePlugin.getLoggerFactory().create(AfkStatusProvider.class);
         this.afkCheckedPlayers = new ConcurrentHashMap<>();
         this.afkPlayerHandling = afkHandling;
 
@@ -31,9 +35,9 @@ public class AfkStatusProvider {
 
     public void reloadConfigValues() {
         afkPlayerHandling.reloadConfigValues();
-        final OptionalLong afkConfigOptional = plugin.getParser().parseToSeconds(plugin.getConfig().getString("afk.after", "15m"));
+        final OptionalLong afkConfigOptional = loriTimePlugin.getParser().parseToSeconds(loriTimePlugin.getConfig().getString("afk.after", "15m"));
         if (afkConfigOptional.isEmpty()) {
-            plugin.getLogger().severe("Can not start the afk-check. No valid afk-config-value present! Skipping the process...");
+            log.error("Can not start the afk-check. No valid afk-config-value present! Skipping the process...");
             return;
         }
         afkConfigTime = afkConfigOptional.getAsLong() * 1000L;
@@ -41,7 +45,7 @@ public class AfkStatusProvider {
 
     public void restartAfkCheck() {
         stopAfkCheck();
-        final boolean afkEnabled = plugin.getConfig().getBoolean("afk.enabled", false);
+        final boolean afkEnabled = loriTimePlugin.getConfig().getBoolean("afk.enabled", false);
         if (afkEnabled) {
             startAfkCheck();
         }
@@ -49,11 +53,11 @@ public class AfkStatusProvider {
 
     private void startAfkCheck() {
         if (afkConfigTime <= 0) {
-            plugin.getLogger().severe("afk.after time needs to be at least 1! Disabling afk feature...");
+            log.error("afk.after time needs to be at least 1! Disabling afk feature...");
             return;
         }
-        final int repeatTime = plugin.getConfig().getInt("afk.repeatCheck", 30);
-        afkCheck = plugin.getScheduler().scheduleAsync(repeatTime / 2L, repeatTime, this::repeatedTimeCheck);
+        final int repeatTime = loriTimePlugin.getConfig().getInt("afk.repeatCheck", 30);
+        afkCheck = loriTimePlugin.getScheduler().scheduleAsync(repeatTime / 2L, repeatTime, this::repeatedTimeCheck);
     }
 
     private void stopAfkCheck() {
@@ -69,7 +73,7 @@ public class AfkStatusProvider {
             final LoriTimePlayer player = entry.getKey();
             final long playerAfkTime = entry.getValue();
             final long currentTime = System.currentTimeMillis();
-            if (plugin.getServer().getPlayer(player.getUniqueId()).isEmpty()) {
+            if (loriTimePlugin.getServer().getPlayer(player.getUniqueId()).isEmpty()) {
                 afkCheckedPlayers.remove(player);
                 continue;
             }