+++ velocity/src/main/java/com/jannik_kuehn/loritimevelocity/listener/TimeAccumulatorVelocityListener.java
@@ -1,6 +1,7 @@
 package com.jannik_kuehn.loritimevelocity.listener;
 
 import com.jannik_kuehn.common.LoriTimePlugin;
+import com.jannik_kuehn.common.api.logger.LoriTimeLogger;
 import com.jannik_kuehn.common.exception.StorageException;
 import com.velocitypowered.api.event.Subscribe;
 import com.velocitypowered.api.event.connection.DisconnectEvent;
@@ -11,8 +12,11 @@ import java.util.UUID;
 public class TimeAccumulatorVelocityListener {
     private final LoriTimePlugin plugin;
 
+    private final LoriTimeLogger log;
+
     public TimeAccumulatorVelocityListener(final LoriTimePlugin plugin) {
         this.plugin = plugin;
+        this.log = plugin.getLoggerFactory().create(TimeAccumulatorVelocityListener.class);
     }
 
     @Subscribe
@@ -23,7 +27,7 @@ public class TimeAccumulatorVelocityListener {
             try {
                 plugin.getTimeStorage().startAccumulating(uuid, now);
             } catch (final StorageException ex) {
-                plugin.getLogger().warning("could not start accumulating online time for player " + uuid, ex);
+                log.warn("could not start accumulating online time for player " + uuid, ex);
             }
         });
     }
@@ -36,7 +40,7 @@ public class TimeAccumulatorVelocityListener {
             try {
                 plugin.getTimeStorage().stopAccumulatingAndSaveOnlineTime(uuid, now);
             } catch (final StorageException ex) {
-                plugin.getLogger().warning("error while stopping accumulation of online time for player " + uuid, ex);
+                log.warn("error while stopping accumulation of online time for player " + uuid, ex);
             }
         });
     }