+++ CHANGELOG.md
@@ -7,8 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased] - ${maven.build.timestamp}
 ### Added
 ### Changed
+- Reworked the plugin logger for better logging
 ### Deprecated
 ### Removed
+- Bukkit compatibility in Order to remove bloated code. Only Paper is supported via the Bukkit loader.
 ### Fixed
 ### Security